Optimize performance of results tree
The GUI is very slow when the number of tests exceeds 500. The reason
for this is the tree component that displays the test names. Its
implementation is inefficient. In each frame, this component draws all
its content - even those invisible to the user.

This commit changes the implementation of the tree component to a more
efficient one. The new implementation only draws the part of the tree
that is currently presented to the user. This component also uses a
minimal amount of RAM, which allows you to run several thousand tests.