Remove assert_same and assert_not_same
These functions have little value and is confusing what they do. It is not obvious how they behave when value is not a reference to a table. Also presenting pointer address to the user is not helpful at all. Whats more assert same implementations are diferent among libraries. For example, Junit's assertSame and Luaunit's assertIs works similarly to Unitron assert_same, but Testify's assert.Same fails when trying to compare strings or numbers.