[bugfix] Picotron crashes when cyclic table is used
When cyclic table is used in message passed to test_fail(), the
Picotron crashes. Picotron has a few bugs related to marshaling messages
sent between processes. One bug is that it crashes when message contains
cyclic table. Another bug is that Picotron is not able to pass very
small numbers (like 0.000001).